Follow the refactoring in metasfresh
ATP in batch entry is wrong after attributes are added https://github.com/metasfresh/metasfresh/issues/4342
